requirements.txt
#ingestion
kafka-python==2.0.2
websocket-client==1.7.0
tenacity==8.2.3
requests==2.31.0
boto3==1.34.0
botocore==1.34.0
rel

# Spark / PySpark — pin khớp image trong spark/Dockerfile (Spark 3.5.8)
pyspark==3.5.8
delta-spark==3.1.0
great-expectations==0.18.19
